requirements.txt
black==21.12b0
flake8==4.0.1
isort==5.10.1
mypy==0.920
pytest==6.2.5
pytest-cov==3.0.0
PyYAML==6.0
types-PyYAML==6.0.3